Tests/LibWeb/Text/expected/css-has-invalidation/has-with-attribute.txt

[set data-matched attribute]
  hasAncestorWalkInvocations: 1
  hasAncestorWalkVisits: 2
  hasMatchInvocations: 1
  hasResultCacheHits: 0
  hasResultCacheMisses: 1
  styleInvalidations: 1
[remove data-matched attribute]
  hasAncestorWalkInvocations: 1
  hasAncestorWalkVisits: 2
  hasMatchInvocations: 1
  hasResultCacheHits: 0
  hasResultCacheMisses: 1
  styleInvalidations: 1
